refactor: Move tests into src folder
