+++ src/__tests__/fixtures/createBannedAttributeTestCases.js
